# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Transfer App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /rest
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /account:
    x-swagger-router-controller: accounts-controller
    get:
      description: Check account exists 
      # used as the method name of the controller
      operationId: check
      parameters:
        - name: agency
          in: query
          description: The value of the bank agency
          required: true
          type: integer
        - name: accountNumber
          in: query
          description: The value of the current bank account
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AccountResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /transference:    
    x-swagger-router-controller: transactions-controller
    post:
      description: Transference amount
      # used as the method name of the controller
      operationId: transference
      consumes:
        - application/json
      produces:
        - application/json
        - text/xml
        - text/html
      parameters:
        - name: body
          in: body
          description: The object of the transference amount
          schema:
            # a pointer to a definition
            $ref: "#/definitions/TransferenceRequest"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/TransferenceResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  AccountResponse:
    required:
      - agency
      - accountNumber
      - cpf
    properties:
      agency:
        type: number
      accountNumber:
        type: number
      cpf:
        type: number
  TransferenceRequest:
    required:
      - sourceAgency
      - sourceAccountNumber
      - destinationAgency
      - destinationAccountNumber
      - value
    properties:
      sourceAgency:
        type: number
      sourceAccountNumber:
        type: number
      destinationAgency:
        type: number
      destinationAccountNumber:
        type: number
      value:
        type: number
  TransferenceResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string